Add/update group target as needed
When user submits a new target, target data table is checked for an
existing entry based on target email. If an existing entry is found
then it is updated, otherwise a new row is added to table. Target
email addresses are now converted to lower case to help prevent
duplicates as well.