Updated Dockerfile and run script
The Dockerfile specifies two separate build containers - one for the
client side JavaScript dependencies, which needs to be minified if
changed, and a build container for Golang.

The Dockerfile has also been rebuilt to use an unprivileged user instead
of root.

The run script adds more options and changes the method the
configuration is overwritten, which may help in situations where the
configuration is not owned by the unprivileged user.