Load datatable rows all at once
The javascripts code which load datatables were drawing the table rows
one at a time. This operation is time intensive and caused the tables
with a couple of thousand (e.g. ~5000 targets associated with a group)
entries to take upto 150 seconds to load. This also caused the pages to
timeout while waiting for data tables to load. This issue has been fixed
by drawing all rows at once.