Use GroupsSummary to create and copy campaign
The Groups (get all groups and associated targets) call is used while
loading modal for creating and copying a campaign. As Groups API gets
all the associated targets for a groups as well, it slows system
considerably if there are large number of groups and targets (~200
groups each with ~100-10000 targets).
As targets are not really needed in this workflow, this call can be
replaced by GroupsSummary call.